Fix issue with OAuth returns not working
Converts the bottom sheet into a separate fragment. This is requires, because as soon as you show any screen on top of a bottom sheet, it will automatically close. So there would be no way to come back to it.
Also there was a hash mismatch between the profile configurations created at 2 places, so I unified the code for that.