#76 Handle server names being too long
It turns out that some of the institutions use a pretty long list of server names, and when joined together, it was longer than 256 characters, which the Enterprise config did not like, and it just reset that field to an empty string, which then made the validity check fail. So we fall back to the old behavior, if we detect that the joined list is too long.